refactor : 로또게임(수동)
내용
* 기존에 Lotteries에서 getter을 제거하기 위해 역할을 위힘했던 rank match 메소드를 dto로 값을 주입을 하는 구조로 변경함에 따라 조금 더 책임에 적합해 보이는 LottoResult클래스로 이관.
* 이관된 메소드에 해당하는 테스트 코드를 이관하고 dto 주입에 관련된 부분을 리팩토링함.
* 컨트롤러에서 값을 일급콜랙션 대신 dto를 주입하도록 수정.
* GenerateLottoNumber에서 불필요하게 사용하던 변수 주입을 제거하고 코드를 리팩토링함.
