+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -7,6 +7,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import lotto.dto.LotteriesDto;
 import lotto.service.LotteryDraw;
 import lotto.service.Rank;
 import org.junit.jupiter.api.DisplayName;
@@ -34,8 +35,21 @@ class LottoResultTest {
         Collections.singletonList(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6))));
     Lotteries lotteries = new Lotteries(testLottos);
 
-    LottoResult lottoResult = LottoResult.getResult(LotteryDraw.createWinLottoInfo("4,5,6,11,12,13", 21), lotteries);
+    LottoResult lottoResult = LottoResult.getResult(LotteryDraw.createWinLottoInfo("4,5,6,11,12,13", 21), LotteriesDto.of(lotteries));
 
     assertThat(lottoResult.allRankResultSum()).isEqualTo(5000);
+  }
+
+  @DisplayName("로또정보들을 기준으로 당첨정보와 매칭하여 결과값을 반환하는 테스트.")
+  @Test
+  void 랭크에맞게매칭값반환() {
+    List<Lotto> lottos = new ArrayList<>();
+    lottos.add(new Lotto(Arrays.asList(1,2,3,4,5,6)));
+
+    Lotteries lotteries = new Lotteries(lottos);
+
+    LottoResult lottoResult = LottoResult.getResult(LotteryDraw.createWinLottoInfo("4,5,6,11,12,13", 21), LotteriesDto.of(lotteries));
+
+    assertThat(lottoResult.getRankCount(Rank.FIFTH)).isEqualTo(1);
   }
 }
