+++ src/test/java/lotto/service/LotteryDrawTest.java
@@ -4,12 +4,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import lotto.domain.Lotteries;
 import lotto.domain.Lotto;
 import lotto.domain.LottoMoney;
 import lotto.domain.LottoResult;
 import lotto.domain.WinLottoInfo;
+import lotto.dto.LotteriesDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -35,10 +37,11 @@ class LotteryDrawTest {
   @DisplayName("수익률 계산 테스트.")
   @Test
   void 당첨정보기준수익률계산() {
-    List<Lotto> testLottos = new ArrayList<>(Arrays.asList(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6))));
+    List<Lotto> testLottos = new ArrayList<>(
+        Collections.singletonList(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6))));
     Lotteries lotteries = new Lotteries(testLottos);
 
-    LottoResult lottoResult = LottoResult.getResult(LotteryDraw.createWinLottoInfo("4,5,6,11,12,13", 21), lotteries);
+    LottoResult lottoResult = LottoResult.getResult(LotteryDraw.createWinLottoInfo("4,5,6,11,12,13", 21), LotteriesDto.of(lotteries));
 
     assertThat(LotteryDraw.gradingScore(lottoResult, new LottoMoney(1000))).isEqualTo(5);
   }
